chore: gate fixes for #8891 (canonical header read, stale pin, ceilings, splits)
