+++ crates/perry-runtime/src/object/mod.rs
@@ -609,8 +609,10 @@ fn keys_index_insert(
 
 pub(crate) mod array_tail_transition;
 mod call_method_depth;
+mod meta_accessors;
 use call_method_depth::CallMethodDepthGuard;
 pub(crate) use call_method_depth::{call_method_depth_restore, call_method_depth_savepoint};
+pub(crate) use meta_accessors::*;
 
 /// Fast direct-mapped inline cache for class shape keys arrays.
 /// Indexed by `shape_id mod CACHE_SIZE`. Each slot stores
@@ -1721,12 +1723,15 @@ const _: () = assert!(std::mem::size_of::<crate::array::ArrayHeader>() == 8);
 /// Promise and Date. Anything else (Temporal, the typed-array views) returns
 /// `None` and keeps its existing storage.
 pub(crate) unsafe fn cell_meta_slot(user_ptr: usize) -> Option<*mut *mut ObjectMeta> {
-    if user_ptr == 0 || user_ptr < 0x10000 || !user_ptr.is_multiple_of(8) {
+    // Canonical validated read rather than an open-coded magnitude test:
+    // `try_read_gc_header` applies `is_plausible_heap_addr` AND rejects
+    // small-buffer slab addresses, which are heap-plausible but carry no
+    // GcHeader — reading one classifies the previous slab entry's bytes as a
+    // type tag.
+    let Some(gc_hdr) = crate::value::addr_class::try_read_gc_header(user_ptr) else {
         return None;
-    }
-    let gc_hdr =
-        (user_ptr as *const u8).sub(crate::gc::GC_HEADER_SIZE) as *const crate::gc::GcHeader;
-    match (*gc_hdr).obj_type {
+    };
+    match gc_hdr.obj_type {
         crate::gc::GC_TYPE_OBJECT => {
             Some(&mut (*(user_ptr as *mut ObjectHeader)).meta as *mut *mut ObjectMeta)
         }
@@ -1765,102 +1770,6 @@ pub(crate) unsafe fn cell_has_meta_edge(user_ptr: usize) -> bool {
 /// The allocation can trigger a collection that MOVES the owner, so the slot
 /// is re-resolved from the rooted address afterwards rather than reusing the
 /// pointer taken before the allocation.
-pub(crate) unsafe fn object_meta_ensure_for_cell(user_ptr: usize) -> Option<*mut ObjectMeta> {
-    let slot = cell_meta_slot(user_ptr)?;
-    if !(*slot).is_null() {
-        return Some(*slot);
-    }
-    let scope = crate::gc::RuntimeHandleScope::new();
-    let owner = scope.root_raw_mut_ptr(user_ptr as *mut u8);
-    let meta = arena_alloc_gc(
-        std::mem::size_of::<ObjectMeta>(),
-        8,
-        crate::gc::GC_TYPE_OBJECT_META,
-    ) as *mut ObjectMeta;
-    let user_ptr = owner.get_raw_mut_ptr::<u8>() as usize;
-    let slot = cell_meta_slot(user_ptr)?;
-    if !(*slot).is_null() {
-        // A re-entrant path installed one while we allocated; keep it.
-        return Some(*slot);
-    }
-    (*meta).prototype = 0;
-    (*meta).attr_key_bits = 0;
-    (*meta).accessor_key_bits = 0;
-    (*meta).flags = 0;
-    (*meta).spill = 0;
-    (*meta).private_evaluation_brand = 0;
-    (*meta).array_subclass_named_prefix_token = 0;
-    (*meta).array_tail_object_hot = 0;
-    (*meta).array_subclass_dense_key = 0;
-    (*meta).array_subclass_dense_slots = 0;
-    (*meta).array_subclass_dense_bounds = 0;
-    (*meta).expando = 0;
-    // GC_STORE_AUDIT(BARRIERED): header-slot store followed by an object-slot
-    // barrier, exactly as `object_meta_ensure` does for an `ObjectHeader`.
-    *slot = meta;
-    crate::gc::runtime_write_barrier_slot(user_ptr, slot as usize, meta as u64);
-    Some(meta)
-}
-
-pub(crate) unsafe fn object_meta_ensure(obj: *mut ObjectHeader) -> *mut ObjectMeta {
-    if !(*obj).meta.is_null() {
-        return (*obj).meta;
-    }
-    // Root the owner across the allocation: `arena_alloc_gc` can trigger a
-    // copied-minor that MOVES `obj`, and the header store below must land
-    // in the live copy, not the stale from-space one. Reload through the
-    // handle after the allocation. (The fresh `meta` record itself cannot
-    // move before the store — no allocation happens in between.)
-    let scope = crate::gc::RuntimeHandleScope::new();
-    let obj_handle = scope.root_raw_mut_ptr(obj);
-    let meta = arena_alloc_gc(
-        std::mem::size_of::<ObjectMeta>(),
-        8,
-        crate::gc::GC_TYPE_OBJECT_META,
-    ) as *mut ObjectMeta;
-    let obj = obj_handle.get_raw_mut_ptr::<ObjectHeader>();
-    if !(*obj).meta.is_null() {
-        // A GC-triggered re-entrant path installed one meanwhile; keep it
-        // (the fresh record above is unreferenced and dies with the cycle).
-        return (*obj).meta;
-    }
-    (*meta).prototype = 0;
-    (*meta).attr_key_bits = 0;
-    (*meta).accessor_key_bits = 0;
-    (*meta).flags = 0;
-    (*meta).spill = 0;
-    (*meta).private_evaluation_brand = 0;
-    (*meta).array_subclass_named_prefix_token = 0;
-    (*meta).array_tail_object_hot = 0;
-    (*meta).array_subclass_dense_key = 0;
-    (*meta).array_subclass_dense_slots = 0;
-    (*meta).array_subclass_dense_bounds = 0;
-    (*meta).expando = 0;
-    // GC_STORE_AUDIT(BARRIERED): meta-record edge is a header-slot store
-    // followed by an object-slot barrier, mirroring `set_object_keys_array`.
-    (*obj).meta = meta;
-    crate::gc::runtime_write_barrier_slot(
-        obj as usize,
-        &(*obj).meta as *const _ as usize,
-        meta as u64,
-    );
-    meta
-}
-
-/// GC slot accessor for the `meta` header edge (#6759 Phase B): a raw-pointer
-/// child slot. The GC type table calls this
-/// only for `GC_TYPE_OBJECT`; RegExp uses its dedicated slot descriptor.
-pub(crate) unsafe fn gc_object_meta_slot(user_ptr: usize) -> Option<*mut u64> {
-    if user_ptr == 0 {
-        return None;
-    }
-    let obj = user_ptr as *mut ObjectHeader;
-    if (*obj).meta.is_null() {
-        return None;
-    }
-    Some(&mut (*obj).meta as *mut _ as *mut u64)
-}
-
 #[inline]
 unsafe fn set_object_keys_array(obj: *mut ObjectHeader, keys_array: *mut ArrayHeader) {
     let live = object_live_slot_count(obj);
